Move away from a dedicated file
